+++ Patches.cs
@@ -1,17 +1,16 @@
 ﻿using BepInEx.Logging;
 using HarmonyLib;
+using System;
 using UnityEngine;
 using static Rats.Plugin;
 
 namespace Rats
 {
-    [HarmonyPatch(typeof(Landmine))]
-    internal class Patches
+    [HarmonyPatch]
+    internal static class Patches
     {
-        private static ManualLogSource logger = Plugin.logger;
-
         [HarmonyPostfix]
-        [HarmonyPatch(nameof(Landmine.SpawnExplosion))]
+        [HarmonyPatch(typeof(Landmine), nameof(Landmine.SpawnExplosion))]
         public static void SpawnExplosionPostfix(Vector3 explosionPosition, bool spawnExplosionEffect = false, float killRange = 1f, float damageRange = 1f, int nonLethalDamage = 50, float physicsForce = 0f, GameObject overridePrefab = null, bool goThroughCar = false)
         {
             if (!IsServerOrHost) { return; }